Move projects array into arraySkills
