+++ arraySkills.js
@@ -225,3 +225,35 @@ export const arraySkills = [
         ]
     }
 ];
+export const arrayProjects = [
+  {
+    titleKey: 'GitZipQR',
+    descKey: 'This script on TypeScript for crypto archive data in QR code, max-security data: all data for decode located straigh in QR code. END-TO-END CRYPTO. N amount passwords. Share your data secure!',
+    link: 'https://github.com/RestlessByte/GitZipQR',
+    icon: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-brand-telegram"><path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M15 10l-4 4l6 6l4 -16l-18 7l4 2l2 6l3 -4" /></svg>`
+  },
+  {
+    titleKey: 'Telegram Game Bot',
+    descKey: 'Interactive game chat bot with real-time engagement features.',
+    link: 'https://t.me/farmilochka0_bot',
+    icon: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-brand-telegram"><path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M15 10l-4 4l6 6l4 -16l-18 7l4 2l2 6l3 -4" /></svg>`
+  },
+  {
+    titleKey: 'Telegram AI Hub Bot [AI integration]',
+    descKey: 'Centralized hub for AI-powered tools and chatbots integration. Create your own chatbots and optimize your response time!',
+    link: 'https://t.me/AIHub0_bot',
+    icon: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-brand-telegram"><path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M15 10l-4 4l6 6l4 -16l-18 7l4 2l2 6l3 -4" /></svg>`
+  },
+  {
+    titleKey: 'AI CLI LINUX',
+    descKey: 'AI integration in CLI LINUX. Meet with: ChatGPT, DeepSeek, Grok, Mistral and other AI models Communicate directlyin TERMINAL [CLI] LINUX, Let\'s give commands by automating our work in LINUX',
+    link: 'https://github.com/RestlessByte/AI-CLI-LINUX',
+    icon: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-brand-telegram"><path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M15 10l-4 4l6 6l4 -16l-18 7l4 2l2 6l3 -4" /></svg>`
+  },
+  {
+    titleKey: 'usingOpenAI',
+    descKey: 'This SDK on TypeScript for integration AI models in your code',
+    link: 'https://github.com/RestlessByte/usingOpenAI',
+    icon: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-brand-telegram"><path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M15 10l-4 4l6 6l4 -16l-18 7l4 2l2 6l3 -4" /></svg>`
+  }
+];
